requirements.txt
pyshark
requests
netifaces
python-dotenv
